+++ app/src/main/java/com/pes/pockles/model/Location.kt
@@ -1,3 +1,3 @@
 package com.pes.pockles.model
 
-data class Location(var latitude: Float, var longitude: Float)
+data class Location(var latitude: Double, var longitude: Double)
